Extract image layer constant loader
